+++ core/src/main/java/com/contentful/vault/Space.java
@@ -47,4 +47,7 @@ public @interface Space {
    * the {@code dbVersion} property.
    */
   String copyPath() default "";
+
+  /** Array of locale codes. */
+  String[] locales();
 }