+++ tests-integration/src/test/java/com/contentful/vaultintegration/lib/allthethings/AllTheThingsSpace.java
@@ -2,6 +2,6 @@ package com.contentful.vaultintegration.lib.allthethings;
 
 import com.contentful.vault.Space;
 
-@Space(value = "hwp8ctxa24x7", models = { AllTheThingsResource.class })
+@Space(value = "hwp8ctxa24x7", models = AllTheThingsResource.class, locales = "en-US")
 public class AllTheThingsSpace {
 }