+++ tests-integration/src/test/java/com/contentful/vaultintegration/lib/links/LinksSpace.java
@@ -2,6 +2,6 @@ package com.contentful.vaultintegration.lib.links;
 
 import com.contentful.vault.Space;
 
-@Space(value = "fp9iccz5ot6h", models = { AssetsContainer.class })
+@Space(value = "fp9iccz5ot6h", models = AssetsContainer.class, locales = "en-US")
 public class LinksSpace {
 }